requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gradio==5.7.1
gradio_client==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.1
safehttpx==0.1.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.12.0
tqdm==4.67.1
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0
wrapt==1.17.0
